+++ maptasker/src/maputil2.py
@@ -159,11 +159,7 @@ def translate_string(text: str, set_language: bool = False) -> str:
             if set_language:
                 lang_to_set = PrimeItems._(text) if text not in PrimeItems.languages else text
                 T.set_language(lang_to_set)
-
-            # One time only, define the length of the "No Profile" translated string.
-            # FIX PrimeItems.no_profile_translated_length on a change in language and not here!!!.
-            sdafadsfc
-            if PrimeItems.no_profile_translated_length == 0:
+                # Set/reset "No Profile" string length in new language.
                 no_profile_text = PrimeItems._("No Profile")
                 PrimeItems.no_profile_translated_length = len(no_profile_text)
             return PrimeItems._(text)